Move tests to root.
This takes advantage of Django 1.6's test runner. We also stop shipping our tests.
