+++ tests/http.py
@@ -1,12 +1,12 @@
 import json
 import random
 
-from ._utils import HatbandTestCase as TestCase
+from ._utils import HatbandTestCase
 
-from ..http import JsonResponse
+from armstrong.hatband.http import JsonResponse
 
 
-class JsonResponseTestCase(TestCase):
+class JsonResponseTestCase(HatbandTestCase):
     def test_turns_body_into_json(self):
         data = {
             "foo": "bar",
